requirements.txt
ultralytics
pandas
numpy
scikit-learn
pillow
